requirements.txt
# shred_net v2 requirements — grouped by pipeline stage

# core cv
opencv-python>=4.9.0
numpy>=1.26.0
scipy>=1.12.0          # pose graph optimization (L-BFGS-B)

# eac-net (feature extraction + training)
torch>=2.2.0
torchvision>=0.17.0

# retrieval
faiss-cpu>=1.8.0       # swap to faiss-gpu on cuda machines

# config
pyyaml>=6.0

# testing
pytest>=8.0.0

# ── future / optional ──────────────────────────────────────────────────────────
torch-geometric>=2.5.0    # if upgrading solver to full GNN
g2o-python                # industrial-grade pose graph optimization
kornia>=0.7.0             # differentiable cv ops (GPU spatial transforms)
hydra-core>=1.3.0         # advanced config composition (replaces pyyaml)